requirements.txt
Django==4.2.2
Pillow==9.5.0
torch==2.0.0

transformers==4.30.1
